Handshake with a hypercore peer on the network
